+++ demo-plugin/src/main/resources/META-INF/plugin.xml
@@ -18,7 +18,7 @@
                          implementationClass="org.jetbrains.research.refactoringDemoPlugin.featureEnvy.MyFeatureEnvyInspection"/>
 
         <toolWindow id="Project statistics" secondary="true" anchor="bottom"
-                    factoryClass="org.jetbrains.research.refactoringDemoPlugin.metrics.MyToolWindowFactory"
+                    factoryClass="org.jetbrains.research.refactoringDemoPlugin.statistics.MyToolWindowFactory"
         />
     </extensions>
